merge

+++ MacrosTracker.Models/DayModels/DayModels.cs
@@ -17,15 +17,15 @@ namespace MacrosTracker.Models.DayModels
     {
         public int DayId { get; set; }
         public DateTime Date { get; set; }
-        public double DayCarbs { get; set; }
+        public int DayCarbs { get; set; }
        
-        public double DayProteins { get; set; }
-        public double DayFats { get; set; }
-        public double DayCalories { get; set; }
-        public double PlusOrMinusCalories { get; set; }
-        public double PlusOrMinusCarbs { get; set; }
-        public double PlusOrMinusProteins { get; set; }
-        public double PlusOrMinusFats { get; set; }
+        public int DayProteins { get; set; }
+        public int DayFats { get; set; }
+        public int DayCalories { get; set; }
+        public int PlusOrMinusCalories { get; set; }
+        public int PlusOrMinusCarbs { get; set; }
+        public int PlusOrMinusProteins { get; set; }
+        public int PlusOrMinusFats { get; set; }
     }
     public class SpecificDayDetail : DayDetail
     {
